public IniLoader class; update type definition
