+++ lib/credentials/shared_ini_file_credentials.js
@@ -1,5 +1,5 @@
 var AWS = require('../core');
-var sharedIniFile = require('../shared-ini');
+var iniLoader = require('../shared-ini').iniLoader;
 var STS = require('../../clients/sts');
 
 /**
@@ -98,12 +98,12 @@ AWS.SharedIniFileCredentials = AWS.util.inherit(AWS.Credentials, {
       var profiles = {};
       var profilesFromConfig = {};
       if (process.env[AWS.util.configOptInEnv]) {
-        var profilesFromConfig = sharedIniFile.loadFrom({
+        var profilesFromConfig = iniLoader.loadFrom({
           isConfig: true,
           filename: process.env[AWS.util.sharedConfigFileEnv]
         });
       }
-      var profilesFromCreds = sharedIniFile.loadFrom({
+      var profilesFromCreds = iniLoader.loadFrom({
         filename: this.filename ||
           (process.env[AWS.util.configOptInEnv] && process.env[AWS.util.sharedCredentialsFileEnv])
       });
