+++ lib/node_loader.js
@@ -36,6 +36,8 @@ AWS.XML.Parser = require('./xml/node_parser');
 // Load Node HTTP client
 require('./http/node');
 
+require('./shared-ini/ini-loader');
+
 // Load custom credential providers
 require('./credentials/ec2_metadata_credentials');
 require('./credentials/remote_credentials');
@@ -87,7 +89,7 @@ AWS.util.update(AWS.Config.prototype.keys, {
         {filename: env[AWS.util.sharedCredentialsFileEnv]},
         {isConfig: true, filename: env[AWS.util.sharedConfigFileEnv]}
       ];
-      var iniLoader = require('./shared-ini');
+      var iniLoader = require('./shared-ini').iniLoader;
       while (!region && toCheck.length) {
         var configFile = iniLoader.loadFrom(toCheck.shift());
         var profile = configFile[env.AWS_PROFILE || AWS.util.defaultProfile];
