+++ lib/publisher/configuration.js
@@ -1,5 +1,5 @@
 var AWS = require('../core');
-var iniLoader = require('../shared-ini');
+var iniLoader = require('../shared-ini').iniLoader;
 /**
  * Resolve client-side monitoring configuration from either environmental variables
  * or shared config file. Configurations from environmental variables have higher priority
