+++ lib/shared-ini/index.js
@@ -1,6 +1,6 @@
-var IniLoader = require('./ini-loader');
+var IniLoader = require('./ini-loader').IniLoader;
 /**
  * Singleton object to load specified config/credentials files.
  * It will cache all the files ever loaded;
  */
-module.exports = new IniLoader();
+module.exports.iniLoader = new IniLoader();
