+++ lib/shared-ini/ini-loader.js
@@ -2,10 +2,21 @@ var AWS = require('../core');
 var os = require('os');
 var path = require('path');
 
-/**
- * @api private
- */
-module.exports = AWS.util.inherit({
+function parseFile(filename, isConfig) {
+  if (!isConfig) {
+    return AWS.util.ini.parse(AWS.util.readFileSync(filename));
+  } else {
+    var content = AWS.util.ini.parse(AWS.util.readFileSync(filename));
+    var tmpContent = {};
+    Object.keys(content).forEach(function(profileName) {
+      var profileContent = content[profileName];
+      tmpContent[profileName.replace(/^profile\s/, '')] = profileContent;
+    })
+    return tmpContent;
+  }
+}
+
+var IniLoader = AWS.util.inherit({
   constructor: function IniLoader() {
     this.resolvedProfiles = {};
   },
@@ -24,19 +35,10 @@ module.exports = AWS.util.inherit({
     return this.resolvedProfiles[filename];
   },
 
-  parseFile: function parseFile(filename, isConfig) {
-    if (!isConfig) {
-      return AWS.util.ini.parse(AWS.util.readFileSync(filename));
-    } else {
-      var content = AWS.util.ini.parse(AWS.util.readFileSync(filename));
-      var tmpContent = {};
-      Object.keys(content).forEach(function(profileName) {
-        var profileContent = content[profileName];
-        tmpContent[profileName.replace(/^profile\s/, '')] = profileContent;
-      })
-      return tmpContent;
-    }
-  },
+  /**
+   * @api private
+   */
+  parseFile: parseFile,
 
   /**
    * @api private
@@ -71,3 +73,10 @@ module.exports = AWS.util.inherit({
     );
   }
 });
+
+AWS.IniLoader = IniLoader;
+
+module.exports = {
+  IniLoader: IniLoader,
+  parseFile: parseFile,
+}
