+++ test/publisher/configuration.spec.js
@@ -2,7 +2,7 @@ var helpers = require('../helpers');
 var AWS = helpers.AWS;
 var spyOn = helpers.spyOn;
 var monitoringConfig = require('../../lib/publisher/configuration');
-var sharedIniFile = require('../../lib/shared-ini')
+var iniLoader = require('../../lib/shared-ini').iniLoader;
 
 if (AWS.util.isNode()) {
   describe('configuration resolving', function () {
@@ -22,7 +22,7 @@ if (AWS.util.isNode()) {
     });
 
     afterEach(function() {
-      sharedIniFile.clearCachedFiles();
+      iniLoader.clearCachedFiles();
     });
   
     describe('get configurations from environmental variable', function () {
